CFP2-158: Split huge layout component into smaller ones
