+++ src/components/main-app/model/toggle-nav-panel.props.ts
@@ -1,0 +1,3 @@
+export type ToggleNavPanelProps = {
+    handleToggleNavPanel: (e: any) => void;
+};
