+++ src/components/tenancy/tenancy-form.tsx
@@ -14,7 +14,7 @@ import {
 } from '@material-ui/core';
 import Autocomplete from '@material-ui/lab/Autocomplete';
 import { months } from '../../constants';
-import { stylingConstants } from '../../theme/shared-styles';
+import { gridSpacing } from '../../theme/shared-styles';
 import { PropertyModel } from '../property/model/property.model';
 import { RentalUnitModel } from '../rental-unit/model/rental-unit.model';
 import FormSubmitBar from '../forms/form-submit-bar/form-submit-bar';
@@ -71,14 +71,14 @@ export const TenancyForm = ({
             component={'form'}
             onSubmit={(e: React.FormEvent<any>) => handleSubmit(e)}
             container
-            spacing={stylingConstants.gridSpacing}
+            spacing={gridSpacing}
         >
             <Grid
                 item
                 container
                 xs={12}
                 sm={6}
-                spacing={stylingConstants.gridSpacing}
+                spacing={gridSpacing}
                 alignItems={'center'}
                 alignContent={'flex-start'}
             >
@@ -253,7 +253,7 @@ export const TenancyForm = ({
                 container
                 xs={12}
                 sm={6}
-                spacing={stylingConstants.gridSpacing}
+                spacing={gridSpacing}
                 alignItems={'center'}
                 alignContent={'flex-start'}
             >
